# backend/Services/UserService/UserService.API/DTOs/Users/CreateUserDto.cs
using System.ComponentModel.DataAnnotations;

namespace UserService.API.DTOs.Users;

public class CreateUserDto
{
    [Required]
    [EmailAddress]
    public string Email { get; set; } = string.Empty;

    [Required]
    [MaxLength(50)]
    public string FirstName { get; set; } = string.Empty;

    [Required]
    [MaxLength(50)]
    public string LastName { get; set; } = string.Empty;
}